Revert "Added cocktails"
