+++ components/CreateRecipie/index.tsx
@@ -1,45 +1,89 @@
-import { BodyGetOpenAiResult } from "@/pages/api/open-ai/food";
-import { TextField, Typography } from "@mui/material";
+import { BodyGetOpenAiResult } from "@/pages/api/open-ai";
+import {
+  Button,
+  Card,
+  FormControlLabel,
+  FormGroup,
+  Grid,
+  Switch,
+  TextField,
+  Typography,
+} from "@mui/material";
 import { Box } from "@mui/system";
 import LoadingButton from "@mui/lab/LoadingButton";
-import { useMemo, useState } from "react";
-import { FormattedMessage, useIntl } from "react-intl";
-import { useSession } from "next-auth/react";
-import { useQuery } from "@tanstack/react-query";
-import { User } from "../../models/User";
-import LoginCta from "./loginCta";
-import {
-  TypeOfFoodButtonsEs,
-  TypeOfFoodButtonsEn,
-  LanguagesEnum,
-} from "@/utils/createRecepie";
-import LoadingScreen from "./loadingScreen";
-import { updateUser } from "@/lib/api/user";
-import { getDalle2Image } from "@/lib/api/open-ai/dalle-2";
-import FoodType from "./foodType";
-import CountMacros from "./countMacros";
-import ExtraActions from "./extraActions";
-import RecipieDetails from "./recipieDetails";
-import BuyTokensCta from "../shared/BuyTokensCta";
-import PageHeader from "../shared/header";
-import { getLanguage } from "../CreateCocktail";
+import { useEffect, useMemo, useState } from "react";
+import { FormattedMessage } from "react-intl";
+import { useRouter } from "next/router";
+import logo from "assets/logoRojo.png";
+import logoWhite from "assets/logoBlanco.png";
 
-const getButtonsLanguage = (shortLocale: string) => {
-  switch (shortLocale) {
-    case "es":
-      return TypeOfFoodButtonsEs;
-    case "en":
-      return TypeOfFoodButtonsEn;
-    default:
-      return TypeOfFoodButtonsEs;
+import Image from "next/image";
+import ShareIcon from "@mui/icons-material/Share";
+import { useSession, signIn } from "next-auth/react";
+import { useQuery } from "@tanstack/react-query";
+import ClearIcon from "@mui/icons-material/Clear";
+import { User } from "../../models/User";
+import { styled } from "@mui/material/styles";
+import getStripe from "@/utils/get-stripe";
+import { Oval } from "react-loader-spinner";
+
+const TypeOfFoodButtonsEn = [
+  { label: "🌮  Mexican", value: "Mexican" },
+  { label: "🥗  Vegan", value: "Vegan" },
+  { label: "🍝  Italian", value: "Italian" },
+  { label: "🍣  Sushi", value: "Sushi" },
+  { label: "🇬🇷  Greek", value: "Greek" },
+  { label: "🇪🇸  Spanish", value: "Spanish" },
+  { label: "🇩🇪  German", value: "German" },
+  { label: "🏮 Chinese", value: "Chinese" },
+  { label: "🥟 Korean", value: "Korean" },
+];
+
+const TypeOfFoodButtonsEs = [
+  { label: "🌮  Mexicana", value: "Mexicana" },
+  { label: "🥗  Vegan", value: "Vegana" },
+  { label: "🍝  Italian", value: "Italiana" },
+  { label: "🍣  Sushi", value: "Sushi" },
+  { label: "🇬🇷  Griega", value: "Griega" },
+  { label: "🇪🇸  Española", value: "Española" },
+  { label: "🇩🇪  Alemana", value: "Alemana" },
+  { label: "🏮 China", value: "China" },
+  { label: "🥟  Coreana", value: "Coreana" },
+];
+
+export enum LanguagesEnum {
+  es = "es",
+  en = "en",
+}
+
+const CardWithGradient = styled(Card)(({ theme }) => ({
+  background: `linear-gradient(90deg, ${theme.palette.primary.main} 0%, ${theme.palette.secondary.main} 100%)`,
+}));
+
+export const redirectToStripe = async () => {
+  const response = await fetch("/api/stripe/checkout_sessions");
+  if (response.status !== 200) {
+    console.error(response.status);
+    return;
   }
+  const data = await response.json();
+
+  const stripe = await getStripe();
+  const { error } = await stripe!.redirectToCheckout({
+    sessionId: data.data.id,
+  });
 };
 
 const fetchUser = (): Promise<{ data: User[] }> =>
   fetch("api/user").then((res) => res.json());
 
 const CreateRecipie = () => {
-  const { data: userData, refetch } = useQuery({
+  const {
+    isLoading,
+    error,
+    data: userData,
+    refetch,
+  } = useQuery({
     queryKey: ["user"],
     queryFn: fetchUser,
     initialData: { data: [] },
@@ -49,9 +93,29 @@ const CreateRecipie = () => {
 
   const isAuthenticated = session.status === "authenticated";
 
-  const intl = useIntl();
-  const shortLocale = intl.locale;
-  const foodTypeButtons = getButtonsLanguage(shortLocale);
+  const { locale } = useRouter();
+  const [shortLocale] = locale ? locale.split("-") : ["en"];
+  const [foodTypeButtons, selectedLanguage] = useMemo(() => {
+    switch (shortLocale) {
+      case "es":
+        return [TypeOfFoodButtonsEs, LanguagesEnum.es];
+      case "en":
+        return [TypeOfFoodButtonsEn, LanguagesEnum.en];
+      default:
+        return [TypeOfFoodButtonsEn, LanguagesEnum.en];
+    }
+  }, [shortLocale]);
+
+  const shareCTAText = useMemo(() => {
+    switch (shortLocale) {
+      case "es":
+        return "\n\n ¿Quieres crear tus propias recetas con AI? \n  https://aifoodie.co";
+      case "en":
+        return "\n\n Want to create your own recipies using AI? \n  https://aifoodie.co";
+      default:
+        return "\n\n Want to create your own recipies using AI? \n  https://aifoodie.co";
+    }
+  }, [shortLocale]);
 
   const [foodType, setFoodType] = useState(foodTypeButtons[0].value);
   const [targetProtein, setTargetProtein] = useState("30");
@@ -62,17 +126,28 @@ const CreateRecipie = () => {
   const [alergies, setAlergies] = useState("");
   const [loading, setLoading] = useState(false);
   const [image, setImage] = useState("");
+
   const [result, setResult] = useState("");
 
   const fetchImage = async (prompt: string) => {
     setImage("");
     if (userData?.data[0].availableTokens === 0) {
-      alert("Favor de comprar");
+      alert("Favro de comprar");
       return;
     }
     if (userData?.data.length) {
-      const response = await getDalle2Image(prompt);
-      setImage(response);
+      const response = await fetch("/api/open-ai/dalle-2", {
+        method: "POST",
+        headers: {
+          "Content-Type": "application/json",
+        },
+        body: JSON.stringify({ prompt: prompt.slice(0, 700) }),
+      });
+
+      const data = await response.json();
+      setImage(data.data);
+
+      console.log(data);
     }
     setLoading(false);
   };
@@ -85,12 +160,17 @@ const CreateRecipie = () => {
       return;
     }
     if (userData?.data.length) {
-      await updateUser({
-        availableTokens: (userData.data[0].availableTokens || 10) - 2,
+      await fetch("/api/user", {
+        method: "PUT",
+        headers: {
+          "Content-Type": "application/json",
+        },
+        body: JSON.stringify({
+          availableTokens: (userData.data[0].availableTokens || 10) - 2,
+        }),
       });
       refetch();
-      // TODO: Refactor this
-      const response = await fetch("/api/open-ai/food", {
+      const response = await fetch("/api/open-ai", {
         method: "POST",
         headers: {
           "Content-Type": "application/json",
@@ -125,7 +205,31 @@ const CreateRecipie = () => {
     setLoading(false);
   };
 
-  if (session.status === "loading") return <LoadingScreen />;
+  if (session.status === "loading")
+    return (
+      <Box
+        sx={{
+          width: "100%",
+          display: "flex",
+          alignItems: "center",
+          justifyContent: "center",
+          height: "100vh",
+        }}
+      >
+        <Oval
+          height={80}
+          width={80}
+          color="#EB1245"
+          wrapperStyle={{}}
+          wrapperClass=""
+          visible={true}
+          ariaLabel="oval-loading"
+          secondaryColor="#EC6314"
+          strokeWidth={2}
+          strokeWidthSecondary={2}
+        />
+      </Box>
+    );
 
   return (
     <Box
@@ -145,39 +249,205 @@ const CreateRecipie = () => {
         }}
       >
         {isAuthenticated && (
-          <PageHeader
-            title={<FormattedMessage id="title" defaultMessage="Recipies AI" />}
-            subTitle={
-              <FormattedMessage id="subTitle" defaultMessage="Recipies AI" />
-            }
-          />
+          <>
+            <Image src={logo} alt="Logo" width={200} />
+            <Typography sx={{ mt: 2 }} variant="h4" component="h1">
+              <FormattedMessage id="title" defaultMessage="Recipies AI" />
+            </Typography>
+            <Typography variant="h5" component="h2">
+              <FormattedMessage
+                id="subtitle"
+                defaultMessage="Create you own recipies powered by AI"
+              />
+            </Typography>{" "}
+          </>
         )}
-        {!isAuthenticated && <LoginCta />}
+        {!isAuthenticated && (
+          <CardWithGradient elevation={12} sx={{ p: 5, width: "100%" }}>
+            <Image src={logoWhite} alt="Logo" width={200} />
+            <Typography
+              color="white"
+              sx={{ mt: 2 }}
+              variant="h4"
+              component="h1"
+            >
+              <FormattedMessage id="title" defaultMessage="Recipies AI" />
+            </Typography>
+            <Typography
+              sx={{ mt: 2 }}
+              color="white"
+              variant="h5"
+              component="h2"
+            >
+              <FormattedMessage
+                id="subtitle"
+                defaultMessage="Create you own recipies powered by AI"
+              />
+            </Typography>
+            <Typography
+              sx={{ mt: 3 }}
+              color="white"
+              variant="h5"
+              component="h2"
+            >
+              <FormattedMessage id="signInCTA" />
+            </Typography>
+            <Button
+              fullWidth
+              sx={{ mt: 2, color: "black", backgroundColor: "white" }}
+              variant="contained"
+              onClick={() => signIn()}
+            >
+              <FormattedMessage id="signIn" />
+            </Button>
+          </CardWithGradient>
+        )}
         <Box
           sx={isAuthenticated ? {} : { opacity: 0.4, pointerEvents: "none" }}
         >
-          <FoodType
-            foodType={foodType}
-            setFoodType={setFoodType}
-            foodTypeButtons={foodTypeButtons}
-          />
+          <Box
+            sx={{
+              mt: 5,
+              display: "flex",
+              flexDirection: "column",
+              gap: 2,
+              width: "100%",
+              maxWidth: "600px",
+              px: 1,
+            }}
+          >
+            <Typography variant="h6" component="h3">
+              <FormattedMessage id="foodType" />: {foodType}
+            </Typography>
+            <Grid
+              direction="row"
+              justifyContent="center"
+              alignItems="center"
+              spacing={2}
+              container
+            >
+              {foodTypeButtons.map((button) => {
+                return (
+                  <Grid key={button.value} item>
+                    {foodType === button.value ? (
+                      <Button
+                        variant="contained"
+                        onClick={() => setFoodType(button.value)}
+                      >
+                        {button.label}
+                      </Button>
+                    ) : (
+                      <Button
+                        variant="outlined"
+                        onClick={() => setFoodType(button.value)}
+                      >
+                        {button.label}
+                      </Button>
+                    )}{" "}
+                  </Grid>
+                );
+              })}
+            </Grid>
+          </Box>
+          <Box
+            sx={{
+              mt: 2,
+              display: "flex",
+              flexDirection: "column",
+              gap: 2,
+              width: "100%",
+              p: 2,
+            }}
+          >
+            <FormGroup>
+              <FormControlLabel
+                control={
+                  <Switch
+                    checked={countMacros}
+                    onChange={(e) => setCountMacros(e.target.checked)}
+                  />
+                }
+                label={<FormattedMessage id="targetMacros" />}
+              />
+            </FormGroup>
+            {countMacros && (
+              <>
+                <TextField
+                  id="protein-textfield"
+                  label={<FormattedMessage id="targetMacrosProtein" />}
+                  variant="outlined"
+                  fullWidth
+                  type="number"
+                  value={targetProtein}
+                  onChange={(e) => setTargetProtein(e.target.value)}
+                  InputLabelProps={{
+                    shrink: true,
+                  }}
+                />
+                <TextField
+                  id="carbs-textfield"
+                  label={<FormattedMessage id="targetMacrosCarbs" />}
+                  variant="outlined"
+                  value={targetCarbs}
+                  onChange={(e) => setTargetCarbs(e.target.value)}
+                  fullWidth
+                  type="number"
+                  InputLabelProps={{
+                    shrink: true,
+                  }}
+                />
+              </>
+            )}
+          </Box>
+          <Box
+            sx={{
+              mt: 2,
+              display: "flex",
+              flexDirection: "column",
+              gap: 2,
+              width: "100%",
+              p: 2,
+            }}
+          >
+            <Typography variant="h5" component="h3">
+              <FormattedMessage id="recipieDetails" />
+            </Typography>
+            <TextField
+              id="outlined-basic"
+              label={<FormattedMessage id="recipieDetailsIngredients" />}
+              onChange={(e) => setPrimaryIngredient(e.target.value)}
+              value={primaryIngredient}
+              variant="outlined"
+              fullWidth
+              InputLabelProps={{
+                shrink: true,
+              }}
+            />
+            <TextField
+              id="outlined-basic"
+              label={<FormattedMessage id="recipieDetailsAlergies" />}
+              variant="outlined"
+              value={alergies}
+              onChange={(e) => setAlergies(e.target.value)}
+              fullWidth
+              InputLabelProps={{
+                shrink: true,
+              }}
+            />
+            <TextField
+              id="outlined-basic"
+              label={<FormattedMessage id="personCount" />}
+              type="number"
+              variant="outlined"
+              value={personCount}
+              onChange={(e) => setPersonCount(e.target.value)}
+              fullWidth
+              InputLabelProps={{
+                shrink: true,
+              }}
+            />
+          </Box>
 
-          <CountMacros
-            countMacros={countMacros}
-            setCountMacros={setCountMacros}
-            targetProtein={targetProtein}
-            setTargetProtein={setTargetProtein}
-            targetCarbs={targetCarbs}
-            setTargetCarbs={setTargetCarbs}
-          />
-          <RecipieDetails
-            setPrimaryIngredient={setPrimaryIngredient}
-            primaryIngredient={primaryIngredient}
-            alergies={alergies}
-            setAlergies={setAlergies}
-            personCount={personCount}
-            setPersonCount={setPersonCount}
-          />
           <LoadingButton
             sx={{ mt: 5 }}
             onClick={() =>
@@ -187,7 +457,7 @@ const CreateRecipie = () => {
                 targetCarbs,
                 primaryIngredient,
                 alergies,
-                selectedLanguage: getLanguage(shortLocale),
+                selectedLanguage,
                 countMacros,
                 personCount,
               })
@@ -200,8 +470,28 @@ const CreateRecipie = () => {
           >
             <FormattedMessage id="generateReciepie" />
           </LoadingButton>
-
-          <ExtraActions result={result} setResult={setResult} />
+          <Box sx={{ display: "flex", my: 2, width: "100%", gap: 2 }}>
+            <Button
+              startIcon={<ClearIcon />}
+              color="secondary"
+              fullWidth
+              variant="outlined"
+              onClick={() => setResult("")}
+            >
+              <FormattedMessage id="erase" />
+            </Button>
+            <Button
+              fullWidth
+              color="secondary"
+              onClick={() =>
+                navigator.clipboard.writeText(`${result} ${shareCTAText}`)
+              }
+              variant="outlined"
+              startIcon={<ShareIcon />}
+            >
+              Share
+            </Button>
+          </Box>
 
           <TextField
             sx={{ width: "100%", mt: 2 }}
@@ -233,7 +523,15 @@ const CreateRecipie = () => {
               {userData.data[0].availableTokens}
             </Typography>
           )}
-          <BuyTokensCta />
+          <Button
+            color="secondary"
+            sx={{ my: 2 }}
+            variant="contained"
+            fullWidth
+            onClick={() => redirectToStripe()}
+          >
+            <FormattedMessage id="buyTokensCTA" />
+          </Button>
         </Box>
       </Box>
     </Box>
