+++ components/navBar/index.tsx
@@ -6,7 +6,7 @@ import Button from "@mui/material/Button";
 import { useSession, signOut } from "next-auth/react";
 import { FormattedMessage } from "react-intl";
 import { styled } from "@mui/material/styles";
-import Link from "next/link";
+import { redirectToStripe } from "../CreateRecipie";
 
 const AppBarWithTheme = styled(AppBar)(({ theme }) => ({
   background: `linear-gradient(90deg, ${theme.palette.primary.main} 0%, ${theme.palette.secondary.main} 100%)`,
@@ -26,19 +26,13 @@ export default function NavBar() {
                 width: "100%",
               }}
             >
-              <Box>
-                <Link style={{ textDecoration: "none" }} href="/">
-                  <Button sx={{ color: "white" }} color="inherit">
-                    <FormattedMessage id="Home" />
-                  </Button>
-                </Link>
-                <Link style={{ textDecoration: "none" }} href="/cocktails">
-                  <Button sx={{ color: "white" }} color="inherit">
-                    <FormattedMessage id="Cocktails" />
-                  </Button>
-                </Link>
-              </Box>
-
+              <Button
+                onClick={() => redirectToStripe()}
+                variant="contained"
+                color="secondary"
+              >
+                <FormattedMessage id="buyTokensCTA" />
+              </Button>
               <Button onClick={() => signOut()} color="inherit">
                 <FormattedMessage id="signOut" />
               </Button>
