--- pages/cocktails.tsx
@@ -1,11 +1,0 @@
-import CreateCocktail from "@/components/CreateCocktail";
-import Seo from "@/components/SEO/Seo";
-
-export default function Home() {
-  return (
-    <>
-      <Seo />
-      <CreateCocktail />
-    </>
-  );
-}
