--- utils/createCocktail.ts
@@ -1,13 +1,0 @@
-export const TypeOfCocktailButtonsEn = [
-  { label: "🍶  Shots", value: "Short Drinks" },
-  { label: "🍹  Drinks", value: "Long Drinks" },
-  { label: "🥃  Aperitif", value: "Aperitif" },
-  { label: "🍸  Digestif", value: "Digestif" },
-];
-
-export const TypeOfCocktailButtonsEs = [
-  { label: "🍶  Shots", value: "Shot" },
-  { label: "🍹  Bebida alchólica", value: "Bebida de mixologia" },
-  { label: "🥃  Aperitivo", value: "Aperitivo" },
-  { label: "🍸  Digestivo", value: "Digestivo" },
-];
