--- utils/createRecepie.ts
@@ -1,28 +1,0 @@
-export const TypeOfFoodButtonsEn = [
-  { label: "🌮  Mexican", value: "Mexican" },
-  { label: "🥗  Vegan", value: "Vegan" },
-  { label: "🍝  Italian", value: "Italian" },
-  { label: "🍣  Sushi", value: "Sushi" },
-  { label: "🇬🇷  Greek", value: "Greek" },
-  { label: "🇪🇸  Spanish", value: "Spanish" },
-  { label: "🇩🇪  German", value: "German" },
-  { label: "🏮 Chinese", value: "Chinese" },
-  { label: "🥟 Korean", value: "Korean" },
-];
-
-export const TypeOfFoodButtonsEs = [
-  { label: "🌮  Mexicana", value: "Mexicana" },
-  { label: "🥗  Vegan", value: "Vegana" },
-  { label: "🍝  Italian", value: "Italiana" },
-  { label: "🍣  Sushi", value: "Sushi" },
-  { label: "🇬🇷  Griega", value: "Griega" },
-  { label: "🇪🇸  Española", value: "Española" },
-  { label: "🇩🇪  Alemana", value: "Alemana" },
-  { label: "🏮 China", value: "China" },
-  { label: "🥟  Coreana", value: "Coreana" },
-];
-
-export enum LanguagesEnum {
-  es = "es",
-  en = "en",
-}
